+++ internal/adapter/themoviedb/model.go
@@ -61,7 +61,7 @@ type TVShowShort struct {
 	Poster       *Image
 	FirstAirDate time.Time
 	VoteAverage  float64
-	VoteCount    int
+	VoteCount    uint32
 	Popularity   float64
 }
 
@@ -72,8 +72,8 @@ type TVShow struct {
 	Genres           []string
 	LastAirDate      time.Time
 	NextEpisodeToAir time.Time
-	NumberOfEpisodes int
-	NumberOfSeasons  int
+	NumberOfEpisodes uint32
+	NumberOfSeasons  uint32
 	OriginCountry    []string
 	Seasons          []Season
 	Status           string
@@ -85,11 +85,11 @@ type TVShow struct {
 type Season struct {
 	ID           uint64
 	AirDate      time.Time
-	EpisodeCount int
+	EpisodeCount uint32
 	Name         string
 	Overview     string
 	Poster       *Image
-	SeasonNumber int
+	SeasonNumber uint8
 	VoteAverage  float64
 }
